requirements.txt
numpy
trimesh
pyyaml
scipy
Pillow
rtree
pyembree
pytest
matplotlib